+++ src/graphql/mod.rs
@@ -1,0 +1,38 @@
+use reqwest::Client;
+use serde_json::{Value, Map};
+
+use crate::utils::get_valid_emails;
+
+type Error = Box<dyn std::error::Error>;
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+pub async fn send_invitation(user_emails: Vec<Value> ) -> Result<()> {
+    const SEND_INVITAION_MUTATION: &str = "mutation inviteMembers($params: InviteMemberInput!) {\n  _invite_members(params: $params) {\n    message\n    __typename\n  }\n}";
+    let client = Client::new();
+    let mut map = Map::new();
+    let mut params_map = Map::new();
+    let mut emails_map = Map::new();
+    let valid_emails = get_valid_emails(user_emails);
+    if valid_emails.len() == 0 {
+        Err("Invalid list uploaded!!")?;
+    }
+    emails_map.insert("emails".to_string(), Value::Array(valid_emails));
+    params_map.insert("params".to_string(), Value::Object(emails_map));
+    map.insert("query".to_string(), Value::String(SEND_INVITAION_MUTATION.to_string()));
+    map.insert("variables".to_string(), Value::Object(params_map));
+    let req = client
+        .post("http://localhost:8080/graphql")
+        .header("x-authorizer-admin-secret", "admin")
+        .json(&map);
+
+    let res = req.send().await?;
+    println!("request status: {}", res.status());
+
+    let body = res.bytes().await?;
+
+    let v = body.to_vec();
+    let s = String::from_utf8_lossy(&v);
+    println!("response: {} ", s);
+
+    Ok(())
+}
